requirements.txt
llama-index==0.7.17
langchain==0.0.266
chainlit==0.6.2
nltk==3.8.1
openai==0.27.8
# psycopg2==2.9.7
psycopg2_binary==2.9.7
pgvector==0.2.1
asyncpg==0.28.0